Move chat panel to its own package
